RavenDB-12677 Refactor client flow so that only ONE forced revision is created + apply in Studio

+++ src/Voron/Data/Tables/Table.cs
@@ -959,7 +959,7 @@ namespace Voron.Data.Tables
             return null;
         }
 
-        public IEnumerable<SeekResult> SeekBackwardFrom(TableSchema.SchemaIndexDef index, Slice prefix, Slice last, int skip)
+        public IEnumerable<SeekResult> SeekBackwardFrom(TableSchema.SchemaIndexDef index, Slice prefix, Slice last, long skip)
         {
             var tree = GetTree(index);
             if (tree == null)
